% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addalpha.R
\name{addalpha}
\alias{addalpha}
\title{addalpha - add transparency info to colorset for plotting}
\usage{
addalpha(colors, alpha = 1)
}
\arguments{
\item{colors}{a hexadecimal string as commonly used in R}

\item{alpha}{a transparency factor:
0 - completely invisible
1 - fully opaque}
}
\value{
a hexadecimal string of length \code{colors}
}
\description{
\code{addalpha} adds transparency
    information to a vector in R color format
}
\examples{
    library(RColorBrewer)
    set2 <- brewer.pal(5, 'Set2')
    set2.light <- addalpha(set2, alpha=0.5)

}
