% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation}
\alias{annotation}
\title{Annotation of the Tung et al (2017) single cell RNA-seq data set}
\format{A data frame with 864 rows and 5 variables:
\describe{
  \item{individual}{patient ID}
  \item{replicate}{replicate ID}
  \item{well}{well ID for the corresponding replicate, each well contains one cell}
  \item{batch}{combination of patient ID and replicate ID}
  \item{sample_id}{combination pf patient ID, replicate ID and well ID}
}}
\source{
\url{https://github.com/jdblischak/singleCellSeq}
}
\usage{
annotation
}
\description{
An annotation dataset containing single cell RNA-seq metadata of three patients
distributed on three replicates each.
}
\keyword{datasets}
