% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_sil.R
\name{batch_sil}
\alias{batch_sil}
\title{batch_sil}
\usage{
batch_sil(pca.data, batch, nPCs = 3)
}
\arguments{
\item{pca.data}{a list as created by \code{prcomp}, \code{batch_sil.R} needs \code{$x}:
the principal components (PCs, correctly: the rotated data)}

\item{batch}{vector with the batch covariate (for each cell)}

\item{nPCs}{the number of principal components to use (default: 3)}
}
\value{
The average silhouette width for all clusters. For batch effect, the smaller the better.
   For biological effect, the larger the better.
}
\description{
Determine batch/bio effect using the silhouette coefficient (adopted from scone):
}
\examples{
   \dontrun{
    pca.data <- prcomp(data, center=TRUE)
    batch.silhouette <- batch_sil(pca.data, batch)
    }
}
