% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molecules}
\alias{molecules}
\title{Summarized count table of the Tung et al (2017) single cell RNA-seq data set}
\format{A data frame with 19027 rows and 864 variables:
\describe{
 Rows correspond to ENSEMBL gene IDs. Columns correspond to cell samples (colnames are equal to sample_id in annotation file).
 }}
\source{
\url{https://github.com/jdblischak/singleCellSeq}
}
\usage{
molecules
}
\description{
A dataset containing single cell RNA-seq count data of three patients
distributed on three replicates each.
}
\keyword{datasets}
