% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quality-single-cells}
\alias{quality-single-cells}
\title{Vector of 'good quality' cells of the Tung et al (2017) single cell RNA-seq data set}
\format{A vector of length 564:
\describe{
 Each entry corresponds to 'good quality' samples (equal to sample_id in annotation file).
 }}
\source{
\url{https://github.com/jdblischak/singleCellSeq}
}
\usage{
quality-single-cells
}
\description{
A vector of sample IDs which cells passed the quality control in the single cell RNA-seq experiment of three patients
distributed on three replicates each.
}
\keyword{datasets}
