% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBET.R
\name{kBET}
\alias{kBET}
\title{kBET - k-nearest neighbour batch effect test}
\usage{
kBET(df, batch, k0 = NULL, knn = NULL, testSize = NULL,
  heuristic = FALSE, stats = 100, alpha = 0.05, addTest = FALSE,
  verbose = TRUE, plot = TRUE)
}
\arguments{
\item{df}{dataset (rows: samples, columns: parameters)}

\item{batch}{batch id for each cell}

\item{k0}{number of nearest neighbours to test on (neighbourhood size)}

\item{knn}{a set of nearest neighbours for each cell (optional)}

\item{testSize}{number of data points to test, (10 percent sample size default)}

\item{heuristic}{compute an optimal neighbourhood size k}

\item{stats}{to create a statistics on batch estimates, evaluate 'stats' subsets}

\item{alpha}{significance level}

\item{addTest}{perform an LRT-approximation to the multinomial test AND a multinomial exact test (if appropriate)}

\item{plot}{if stats > 10, then a boxplot of the resulting rejection rates is created}
}
\value{
List with test summary - a rejection rate for the data, an expected rejection rate for random labeling and the significance for the observed result
}
\description{
kBET runs a chi square test to evaluate the probability of a batch effect.
}
\examples{
batch.estimate <- kBET(data,batch)
}
