% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcRegression.R
\name{pcRegression}
\alias{pcRegression}
\title{pcRegression}
\usage{
pcRegression(pca.data, batch, tol = 1e-16)
}
\arguments{
\item{pca.data}{a list as created by 'prcomp', pcRegression needs}

\item{batch}{vector with the batch covariate (for each cell)}

\item{tol}{truncation threshold for significance level, default: 1e-16}
}
\description{
pcRegression does a linear model fit of principal components and a batch (categorical) variable
}
